+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - refactored item flags, see README.md
 - add default params for items
 - add type checks for given params
+- removed recursive deletion in safeDown(). These migrations cannot be reverted. 
 - deprecate _exists and _force item params
 
 ## 1.0.0 under development